moving Person model to core module
